Update session form files
The session form is built using a base form file and login/signup
input components. The base renders the inputs based on the
form type.

I changed the name of the files to correctly represent what they
contain. I also refactored the login_form and signup_form
components to only return inputs and labels. Then I renamed them
to login_inputs and signup_inputs. When they were named forms,
it was bothering me that the file naming did not represent the
heirarchy of the components.