Refactor session form
The session form render function was handling what type of inputs
to render based on the form type. I moved the logic out so that
the inputs are passed in as a component and only dealt with once.
The inputs are now passed in through containers as a component
instead.

This change allowed me to move the session form component as a top
level file in /components/auth. I also moved the input components
into their own folder.