Change Session to Auth and fix AuthInput bug
AuthInput onChange was not updating the state of the of the
AuthForm because the update function passed into the inputs
from the AuthForm was being used incorrectly inside the onChange
function of each AuthInput. It was being called function style
inside the onChange callback but it was returning a function
and not updating the state of the session form. This is because
the update function was originally created to be the only function
passed into onChange, invoked.

All words that were once session were changed to auth. The routes
for auth had to be manually constructed to match the desired
changes