Add closing functionality to Toasts
Through the store connecting container, I pass in a function
that takes in two arguments: type and id. The type is used to
determine what function to return as the close action. The id
is used to dispatch a removal of the corresponding type
for an element with the given id.