Correct Transaction#create current user security
I was checking to see if the current use was involved in the
posted transaction by checking if their id matched the new
transaction's payer_id or the transactee's id. The latter is
incorrect because that will not be the case assuming the data
has not been tampered with. Also, it does not correctly check
for involvement in the transaction. This was updated to check
if the current user's id matches the payer_id or the payee_id.