Fix Toaster CSS and HTML
Fixed the Toaster and dynamically update its top based on the
current path

I realized that the toaster-container was taking up the whole
screen and covering all other components, not allowing input.
I deleted the container and left only the toaster. The CSS
was also reduced to a few lines