Update selections on search selection
I used shouldComponentUpdate to check for a selection. If a
selection is detected, the selections key is updated. Then,
an action to remove the current selection is dispatched. The
component is not updated when a selection is found because it
will be updated again after the current search selection is
removed. Then, there is no selection, so shouldComponentUpdate
returns true.